requirements.txt
python-telegram-bot==20.3
openai
apify-client
python-dotenv
requests
httpx